+++ piecrust/tests/vector.rs
@@ -17,7 +17,7 @@ pub fn vector_push_pop() -> Result<(), Error> {
 
     let id = session.deploy(
         contract_bytecode!("vector"),
-        ContractData::builder(OWNER),
+        ContractData::builder().owner(OWNER),
         LIMIT,
     )?;